refactor: Move remaining content into `data.ts` file
- Also load fonts on mount with `loadCustomFont` instead of `loadCustomFonts`.
