+++ mobile/src/modules/customization/font/utils.ts
@@ -62,14 +62,3 @@ export async function loadCustomFont(fontUri: string) {
   if (!fileName) throw new Error("File name cannot be derived.");
   return loadFontsAsync({ [removeFileExtension(fileName)]: fontUri });
 }
-
-export function loadCustomFonts(fonts: Array<{ uri: string }>) {
-  const customFontEntries = fonts
-    .map((font) => {
-      const fileName = font.uri.split("/").at(-1);
-      if (!fileName) return undefined;
-      return [removeFileExtension(fileName), font.uri] as const;
-    })
-    .filter((entry) => entry !== undefined);
-  return loadFontsAsync(Object.fromEntries(customFontEntries));
-}
