+++ mobile/src/stores/Preference/store.ts
@@ -7,8 +7,8 @@ import { LANGUAGES } from "~/modules/i18n/constants";
 
 import { throwIfNoResults } from "~/lib/drizzle";
 import { createPersistedStore } from "~/lib/zustand";
-import { getCustomFonts } from "~/modules/customization/font/queries";
-import { loadCustomFonts } from "~/modules/customization/font/utils";
+import { getCustomFonts } from "~/modules/customization/font/core/data";
+import { loadCustomFont } from "~/modules/customization/font/utils";
 import { getCustomTheme, resolveCustomTheme } from "~/modules/theme/utils";
 import type { PreferenceStore } from "./constants";
 import { OmittedFields } from "./constants";
@@ -19,12 +19,8 @@ export const preferenceStore = createPersistedStore<PreferenceStore>(
     _hasHydrated: false,
     _init: async (state) => {
       // Load custom fonts.
-      try {
-        const customFonts = await getCustomFonts();
-        await loadCustomFonts(customFonts);
-      } catch (err) {
-        console.log("[Preference Store] Failed to load custom fonts:", err);
-      }
+      const customFonts = await getCustomFonts();
+      await Promise.allSettled(customFonts.map((f) => loadCustomFont(f.uri)));
 
       // Set app theme on initialization.
       try {
